Move whole implementation to a separate directory
