+++ examples/workqueue/worker.go
@@ -3,113 +3,16 @@ package main
 import (
 	"context"
 	"encoding/json"
-	"errors"
 	"fmt"
-	"io/ioutil"
 	"log"
 	"net/url"
 	"os"
 	"os/exec"
 	"path/filepath"
-	"strconv"
 	"time"
 
-	"github.com/nats-io/jsm.go"
-	natsd "github.com/nats-io/nats-server/v2/server"
 	"github.com/nats-io/nats.go"
 )
-
-func main() {
-	natsURL, err := url.Parse(os.Getenv("NATS_URL"))
-	if err != nil {
-		panic(err)
-	}
-	outDir := os.Getenv("OUTDIR")
-
-	srv, nc, err := StartJSServer(natsURL)
-	if err != nil {
-		panic(err)
-	}
-	defer srv.Shutdown()
-	defer nc.Flush()
-
-	stream, err := AddStream("IMAGES", nc)
-	if err != nil {
-		panic(err)
-	}
-	consumer, err := AddConsumer("BW", "IMAGES.blackandwhite", stream.Name(), nc)
-	if err != nil {
-		panic(err)
-	}
-
-	if err = sendMessage(consumer.FilterSubject(), nc); err != nil {
-		panic(err)
-	}
-
-	if err = processNextMessage(stream.Name(), consumer.DurableName(), outDir, nc); err != nil {
-		panic(err)
-	}
-}
-
-func StartJSServer(url *url.URL) (*natsd.Server, *nats.Conn, error) {
-	dir, err := ioutil.TempDir("", "nats-jetstream-*")
-	if err != nil {
-		return nil, nil, err
-	}
-
-	natsPort, err := strconv.Atoi(url.Port())
-	if err != nil {
-		return nil, nil, err
-	}
-
-	opts := &natsd.Options{
-		JetStream: true,
-		StoreDir:  dir,
-		Host:      url.Hostname(),
-		Port:      natsPort,
-		LogFile:   "/dev/stdout",
-		Trace:     true,
-	}
-	s, err := natsd.NewServer(opts)
-	if err != nil {
-		return nil, nil, err
-	}
-	go s.Start()
-	if !s.ReadyForConnections(10 * time.Second) {
-		return nil, nil, errors.New("nats server didn't start")
-	}
-
-	nc, err := nats.Connect(s.ClientURL())
-	if err != nil {
-		return nil, nil, err
-	}
-
-	return s, nc, nil
-}
-
-func AddStream(name string, nc *nats.Conn) (*jsm.Stream, error) {
-	stream, err := jsm.NewStreamFromDefault(name, jsm.DefaultWorkQueue, jsm.StreamConnection(
-		jsm.WithConnection(nc)), jsm.Subjects(name+".*"), jsm.FileStorage(), jsm.MaxAge(24*365*time.Hour),
-		jsm.DiscardOld(), jsm.MaxMessages(-1), jsm.MaxBytes(-1), jsm.MaxMessageSize(512),
-		jsm.DuplicateWindow(1*time.Hour))
-	if err != nil {
-		return nil, err
-	}
-
-	return stream, nil
-}
-
-func AddConsumer(name, filter, stream string, nc *nats.Conn) (*jsm.Consumer, error) {
-	consumer, err := jsm.NewConsumerFromDefault(stream, jsm.DefaultConsumer, jsm.ConsumerConnection(
-		jsm.WithConnection(nc)), jsm.DurableName(name), jsm.MaxDeliveryAttempts(5),
-		jsm.AckWait(30*time.Second), jsm.AcknowledgeExplicit(), jsm.ReplayInstantly(),
-		jsm.DeliverAllAvailable(), jsm.FilterStreamBySubject(filter))
-	if err != nil {
-		return nil, err
-	}
-
-	return consumer, nil
-}
 
 // Job is the request to convert an image
 type Job struct {
@@ -123,31 +26,54 @@ type Advisory struct {
 	ID string `json:"id"`
 }
 
-func sendMessage(subj string, nc *nats.Conn) error {
-	job := Job{
-		ID:       "123456",
-		URI:      "/tmp/images/color.jpg",
-		Advisory: "",
+func panicIfError(err error) {
+	if err == nil {
+		return
 	}
-	data, err := json.Marshal(job)
+
+	panic(err)
+}
+
+func convertMessage(job *Job, outDir string) error {
+	uri, err := url.Parse(job.URI)
+	panicIfError(err)
+
+	// timeout the command after 30 seconds
+	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
+	defer cancel()
+
+	outfile := filepath.Join(outDir, fmt.Sprintf("%s%s", job.ID, filepath.Ext(uri.Path)))
+	log.Printf("Converting %s into %s", uri.Path, outfile)
+
+	// use ImageMagick to convert the image
+	convert := exec.CommandContext(ctx, "/bin/convert", "-monochrome", uri.Path, outfile)
+	return convert.Run()
+}
+
+func publishAdvisory(id string, target string, nc *nats.Conn) error {
+	// Publish an advisory to IMAGES.advisory
+	advisory := &Advisory{id}
+	advj, err := json.Marshal(advisory)
 	if err != nil {
 		return err
 	}
-	if err = nc.Publish(subj, data); err != nil {
-		return err
-	}
 
-	return nil
+	return nc.Publish(target, advj)
 }
 
-func processNextMessage(stream, consumer, outDir string, nc *nats.Conn) error {
-	msg, err := nc.Request(fmt.Sprintf("$JS.API.CONSUMER.MSG.NEXT.%s.%s", stream, consumer), []byte(""), time.Minute)
+func processNextMessage(nc *nats.Conn, outDir string) error {
+	log.Printf("Looking for a B&W conversion job")
+
+	// Load the next message from the IMAGES stream BW consumer
+	msg, err := nc.Request("$JS.API.CONSUMER.MSG.NEXT.IMAGES.BW", []byte(""), time.Minute)
+	// ErrTimeout means there is no new message now lets just skip
 	if err == nats.ErrTimeout {
 		return nil
 	}
 	if err != nil {
 		return err
 	}
+
 	// Parse the job
 	job := &Job{}
 	err = json.Unmarshal(msg.Data, job)
@@ -175,31 +101,19 @@ func processNextMessage(stream, consumer, outDir string, nc *nats.Conn) error {
 	return msg.Respond(nil)
 }
 
-func convertMessage(job *Job, outDir string) error {
-	uri, err := url.Parse(job.URI)
-	if err != nil {
-		return err
-	}
-	// timeout the command after 30 seconds
-	ctx, cancel := context.WithTimeout(context.Background(), 30*time.Second)
-	defer cancel()
+func main() {
+	nc, err := nats.Connect(os.Getenv("NATS_URL"), nats.ErrorHandler(func(_ *nats.Conn, _ *nats.Subscription, err error) {
+		panic(err)
+	}))
+	panicIfError(err)
 
-	outfile := filepath.Join(outDir, fmt.Sprintf("%s%s", job.ID, filepath.Ext(uri.Path)))
-	log.Printf("Converting %s into %s", uri.Path, outfile)
-
-	// use ImageMagick to convert the image
-
-	convert := exec.CommandContext(ctx, "/usr/bin/convert", "-monochrome", uri.Path, outfile)
-	return convert.Run()
-}
-
-func publishAdvisory(id string, target string, nc *nats.Conn) error {
-	// Publish an advisory to IMAGES.advisory
-	advisory := &Advisory{id}
-	advj, err := json.Marshal(advisory)
-	if err != nil {
-		return err
+	outDir := os.Getenv("OUTDIR")
+	if outDir == "" {
+		panic("Please set OUTDIR")
 	}
 
-	return nc.Publish(target, advj)
+	for {
+		err = processNextMessage(nc, outDir)
+		panicIfError(err)
+	}
 }
